fix(playback): break dependency-cruiser cycles for CI
Extract gapless progress tracking and track DTO mapping into leaf modules
so seekAction and index-first enrich paths no longer route through playerStore.
Enrich uses index-only reads; gapless preload refreshes loudness in-store.
